+++ composeApp/src/commonMain/kotlin/ai/rever/boss/components/plugin/language/EditorLanguages.kt
@@ -1,5 +1,7 @@
 package ai.rever.boss.components.plugin.language
 
+import ai.rever.boss.plugin.language.LanguageIds
+
 /**
  * The host's single answer to "what language is this file?".
  *
@@ -13,41 +15,32 @@ package ai.rever.boss.components.plugin.language
  * provider and the built-in editor tab — so `editor_detect_language` cannot report
  * one language while that tab highlights another.
  *
- * It is **not** yet every map in the tree. `EditorServiceImpl` in
- * `modules/boss-app-editor` keeps its own (`plaintext` for unknown, `shell` rather
- * than `bash`, and a `proto` this table lacks), and `FileIcons.forSpecialFileName`
- * in `plugin-platform/plugin-icons` is a third. Neither can depend on
- * `composeApp/commonMain`, so absorbing them means moving this table into a
- * `plugin-platform` module — which both of those, and plugins, *can* consume.
- * Until then those remain hand-maintained.
+ * The table itself now lives in `plugin-platform/plugin-language-types`'s
+ * [LanguageIds] (BossConsole#75): `modules/boss-app-editor`'s `EditorServiceImpl`
+ * could not depend on anything in `composeApp`, so it kept its own smaller,
+ * independently hand-maintained copy that disagreed on `.sh`/`.bash`/`.zsh` (`bash`
+ * here, `shell` there). Both now read the same table. This object stays as a thin
+ * alias so every existing call site (`EditorContentProviderImpl`,
+ * `FileTypeCategoriesTest`) keeps working unchanged.
  *
- * The editor-tab plugin keeps its own `LanguageDetection` (separate artifact, cannot
- * depend on this). Measured against it: the 15 ids added here are identical
- * (`properties` not `ini`, `batch` not `bat`); `sh`/`bash`/`zsh` differ — `bash`
- * here, `shell` there — but its lexer factory accepts both, so they resolve to the
- * same lexer; and `r` has no lexer on either side, so `.r` files are named but not
- * highlighted. Nothing enforces this at build time.
+ * Two further copies are known and are **not** consolidated here - see [LanguageIds]'s
+ * own doc for why: the editor-tab plugin's `LanguageDetection` and BossEditor's lexer
+ * registry both live in separate repositories this build cannot reach, and
+ * `FileIcons.forSpecialFileName`/`forFile` keys icon selection off the raw filename
+ * for a finer granularity than a language id carries, so there is nothing there to
+ * route through this table without a separate redesign.
  *
- * The tables are data rather than a `when` chain — a 40-branch `when` is both harder
- * to scan and, at complexity 44, something detekt rightly objects to.
+ * Measured against the editor-tab plugin's copy: the 15 ids added here are identical
+ * (`properties` not `ini`, `batch` not `bat`); `r` has no lexer on either side, so
+ * `.r` files are named but not highlighted. Nothing enforces this at build time.
  */
 object EditorLanguages {
     /**
      * Language id for [filePath], or `"text"` when nothing matches.
      *
-     * File *name* patterns are tried before the extension, because the files that
-     * most need identifying have no extension: `substringAfterLast('.')` yields `""`
-     * for both `Dockerfile` and `Makefile`.
-     *
-     * The extension is read from the file name, not the whole path. Reading it from
-     * the path let a dot in a parent directory leak into the answer —
-     * `/srv/v1.2/Makefile` produced the "extension" `2/Makefile`.
+     * See [LanguageIds.detect] for the matching rules.
      */
-    fun detect(filePath: String): String {
-        val fileName = filePath.substringAfterLast('/').substringAfterLast('\\')
-        forFileName(fileName)?.let { return it }
-        return EXTENSIONS[fileName.substringAfterLast('.', "").lowercase()] ?: TEXT
-    }
+    fun detect(filePath: String): String = LanguageIds.detect(filePath)
 
     /**
      * The extension-to-language table, for whoever has to stay in step with it.
@@ -59,138 +52,6 @@ object EditorLanguages {
      * means a language it *can* highlight that the OS never sends it.
      * `FileTypeCategoriesTest` compares the two, which is the enforcement this
      * class's own KDoc says nothing has.
-     *
-     * Read-only by construction: `EXTENSIONS` is an immutable `Map`, so this
-     * hands out no way to change it.
      */
-    fun extensions(): Map<String, String> = EXTENSIONS
-
-    /**
-     * Languages identified by file name rather than extension.
-     *
-     * Checked before extensions so that `Dockerfile.dev` is a Dockerfile rather than
-     * whatever `.dev` might otherwise suggest.
-     *
-     * Deliberately omits `CMakeLists.txt`: CMake is not Make, and a Make lexer would
-     * hunt for tab-indented recipes that do not exist while missing `if()/endif()`
-     * and `${VAR}`. No highlighting beats confidently wrong highlighting.
-     */
-    private fun forFileName(fileName: String): String? {
-        val lower = fileName.lowercase()
-        EXACT_NAMES[lower]?.let { return it }
-        return PREFIXED_NAMES.firstNotNullOfOrNull { (prefix, language) ->
-            language.takeIf { lower.startsWith(prefix) }
-        }
-    }
-
-    private const val TEXT = "text"
-
-    /** Whole file names. Bare only — `Gemfile.lock` is generated data, not Ruby. */
-    private val EXACT_NAMES =
-        mapOf(
-            "dockerfile" to "dockerfile",
-            "containerfile" to "dockerfile",
-            "makefile" to "makefile",
-            "gnumakefile" to "makefile",
-            ".env" to "properties",
-            "gemfile" to "ruby",
-            "rakefile" to "ruby",
-        )
-
-    /** Name prefixes, for the `Dockerfile.dev` / `.env.local` family. */
-    private val PREFIXED_NAMES =
-        listOf(
-            "dockerfile." to "dockerfile",
-            "containerfile." to "dockerfile",
-            "makefile." to "makefile",
-            ".env." to "properties",
-        )
-
-    private val EXTENSIONS =
-        mapOf(
-            "kt" to "kotlin",
-            "kts" to "kotlin",
-            "java" to "java",
-            "js" to "javascript",
-            "jsx" to "javascript",
-            "mjs" to "javascript",
-            "cjs" to "javascript",
-            "ts" to "typescript",
-            "tsx" to "typescript",
-            "py" to "python",
-            "pyw" to "python",
-            "json" to "json",
-            "xml" to "xml",
-            "html" to "html",
-            "htm" to "html",
-            "css" to "css",
-            "scss" to "css",
-            "sass" to "css",
-            "md" to "markdown",
-            "markdown" to "markdown",
-            "toml" to "toml",
-            "gradle" to "groovy",
-            "swift" to "swift",
-            "c" to "c",
-            "h" to "c",
-            "cpp" to "cpp",
-            "cc" to "cpp",
-            "cxx" to "cpp",
-            "hpp" to "cpp",
-            "cs" to "csharp",
-            "rs" to "rust",
-            "go" to "go",
-            "rb" to "ruby",
-            "php" to "php",
-            "pl" to "perl",
-            "pm" to "perl",
-            "lua" to "lua",
-            "sh" to "bash",
-            "bash" to "bash",
-            "zsh" to "bash",
-            "yml" to "yaml",
-            "yaml" to "yaml",
-            "sql" to "sql",
-            "r" to "r",
-            "scala" to "scala",
-            // Languages the editor has lexers for that no host map ever named.
-            "dockerfile" to "dockerfile",
-            "mk" to "makefile",
-            "mak" to "makefile",
-            "properties" to "properties",
-            "ini" to "properties",
-            "cfg" to "properties",
-            "env" to "properties",
-            "diff" to "diff",
-            "patch" to "diff",
-            "bat" to "batch",
-            "cmd" to "batch",
-            "clj" to "clojure",
-            "cljs" to "clojure",
-            "cljc" to "clojure",
-            "edn" to "clojure",
-            "tex" to "latex",
-            "sty" to "latex",
-            "cls" to "latex",
-            "bib" to "latex",
-            "lisp" to "lisp",
-            "lsp" to "lisp",
-            "el" to "lisp",
-            "scm" to "lisp",
-            "tcl" to "tcl",
-            "f" to "fortran",
-            "f90" to "fortran",
-            "f95" to "fortran",
-            "f03" to "fortran",
-            "for" to "fortran",
-            "d" to "d",
-            "pas" to "delphi",
-            "dpr" to "delphi",
-            "dfm" to "delphi",
-            "vb" to "visualbasic",
-            "vbs" to "visualbasic",
-            "as" to "actionscript",
-            "jsp" to "jsp",
-            "jspx" to "jsp",
-        )
+    fun extensions(): Map<String, String> = LanguageIds.extensions()
 }
