+++ modules/boss-app-editor/src/test/kotlin/ai/rever/boss/app/editor/EditorServiceImplTest.kt
@@ -1,0 +1,66 @@
+package ai.rever.boss.app.editor
+
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+/**
+ * BossConsole#75: this used to be a hand-maintained table independent of
+ * `composeApp`'s `EditorLanguages`, and the two disagreed on `.sh`/`.bash`/`.zsh`
+ * (`shell` here, `bash` there). Both now read the shared `LanguageIds` table -
+ * `detectLanguage("sh")` returning `bash` rather than `shell` is the actual bug this
+ * consolidation fixes, not just a refactor with no observable effect.
+ */
+class EditorServiceImplTest {
+    private val service = EditorServiceImpl()
+
+    @Test
+    fun `shell extensions now agree with the shared table, not the old local one`() {
+        assertEquals("bash", service.detectLanguage("sh"))
+        assertEquals("bash", service.detectLanguage("bash"))
+        assertEquals("bash", service.detectLanguage("zsh"))
+    }
+
+    @Test
+    fun `previously working mappings are unchanged`() {
+        assertEquals("kotlin", service.detectLanguage("kt"))
+        assertEquals("kotlin", service.detectLanguage("kts"))
+        assertEquals("java", service.detectLanguage("java"))
+        assertEquals("python", service.detectLanguage("py"))
+        assertEquals("javascript", service.detectLanguage("js"))
+        assertEquals("go", service.detectLanguage("go"))
+        assertEquals("rust", service.detectLanguage("rs"))
+        assertEquals("yaml", service.detectLanguage("yaml"))
+        assertEquals("json", service.detectLanguage("json"))
+    }
+
+    @Test
+    fun `proto keeps its local mapping - the shared table has no id for it`() {
+        // LanguageIds is shared with boss-file-types.json's default-app extension
+        // list; adding "proto" there is a separate change, so it stays a local
+        // addition on top of the shared table rather than migrated into it.
+        assertEquals("protobuf", service.detectLanguage("proto"))
+    }
+
+    @Test
+    fun `an unrecognised extension is plaintext, not the shared table's text`() {
+        // EditorServiceImpl's own default was always "plaintext", distinct from
+        // EditorLanguages' "text" - preserved deliberately, since this is this
+        // service's own gRPC contract, not a value composeApp reads.
+        assertEquals("plaintext", service.detectLanguage("notarealextension"))
+    }
+
+    @Test
+    fun `newly available ids the old local table never had`() {
+        // Gained for free by reading the shared table instead of a copy that only
+        // ever knew ~24 extensions.
+        assertEquals("fortran", service.detectLanguage("f90"))
+        assertEquals("clojure", service.detectLanguage("clj"))
+        assertEquals("batch", service.detectLanguage("bat"))
+        assertEquals("diff", service.detectLanguage("diff"))
+    }
+
+    @Test
+    fun `extension lookup is case-insensitive`() {
+        assertEquals("kotlin", service.detectLanguage("KT"))
+    }
+}
