+++ plugin-platform/plugin-language-types/build.gradle.kts
@@ -1,0 +1,45 @@
+import org.jetbrains.kotlin.gradle.dsl.JvmTarget
+
+plugins {
+    kotlin("multiplatform")
+}
+
+group = "ai.rever.boss.plugin"
+
+kotlin {
+    compilerOptions {
+        freeCompilerArgs.add("-Xexpect-actual-classes")
+    }
+    jvmToolchain(17)
+
+    jvm("desktop") {
+        compilations.all {
+            compileTaskProvider.configure {
+                compilerOptions { jvmTarget.set(JvmTarget.JVM_17) }
+            }
+        }
+    }
+
+    sourceSets {
+        commonMain {
+            dependencies {
+                // No dependencies - pure data types
+            }
+        }
+        named("desktopMain") {
+            dependencies {
+                // No dependencies
+            }
+        }
+        named("desktopTest") {
+            dependencies {
+                implementation(kotlin("test-junit5"))
+                implementation(libs.junit.jupiter)
+            }
+        }
+    }
+}
+
+tasks.withType<Test> {
+    useJUnitPlatform()
+}
