+++ settings.gradle.kts
@@ -84,6 +84,7 @@ include(":plugin-platform:plugin-run-types")
 include(":plugin-platform:plugin-workspace-types")
 include(":plugin-platform:plugin-bookmark-types")
 include(":plugin-platform:plugin-icons")
+include(":plugin-platform:plugin-language-types")
 include(":plugin-platform:plugin-path-utils")
 include(":plugin-platform:plugin-sandbox")
 include(":plugin-platform:plugin-api-browser")
